update DECIMAL(19,2) di PostgreSQL data type balance

+++ src/main/java/com/nutech/digitalservice/repository/BalanceRepositoryCustom.java
@@ -3,10 +3,11 @@ package com.nutech.digitalservice.repository;
 import com.nutech.digitalservice.entity.Balance;
 import com.nutech.digitalservice.entity.User;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface BalanceRepositoryCustom {
     Optional<Balance> findByUserWithRawQuery(User user);
-    Balance updateBalanceWithRawQuery(User user, Long newBalance);
-    Balance insertBalanceWithRawQuery(User user, Long initialBalance);
+    Balance updateBalanceWithRawQuery(User user, BigDecimal newBalance);
+    Balance insertBalanceWithRawQuery(User user, BigDecimal initialBalance);
 }
